Separated detailview and detailviewmodel, moved mainactivity
